Rename quantity picker component following naming convention
